Address review and fix bugs
