+++ Algorithm.Python/RiskParityPortfolioAlgorithm.py
@@ -19,7 +19,7 @@ class RiakParityPortfolioAlgorithm(QCAlgorithm):
 
     def Initialize(self):
         self.SetStartDate(2021, 2, 21)  # Set Start Date
-        self.SetEndDate(2021, 3, 31)
+        self.SetEndDate(2021, 3, 30)
         self.SetCash(100000)  # Set Strategy Cash
         self.SetSecurityInitializer(lambda security: security.SetMarketPrice(self.GetLastKnownPrice(security)))
